backend/requirements.txt
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg[binary]==3.2.13; python_version >= "3.13"
whitenoise==6.6.0
django-storages==1.14.4
boto3==1.34.34
pillow==10.2.0; python_version < "3.13"
pillow>=10.4.0; python_version >= "3.13"
